add example3 for programming arduino with assembly
here is a hello world program that set ON the
arduino LED (p13 pin) with the AVR assembly language.

This program require avra and avrdude which can be installed on ubuntu/debian:
$ sudo apt-get install avrdude avra

to build the hello.hex file just do in ex3-led-assembly
$ make

To flash the program on the arduino board.
We do the assumption that arduino is attached to ttyACM0.
you can check your good tty by performing (dmesg | tail) when pluging arduino.
$ make flash

To clean the build:
$ make clean

Signed-off-by: Zakaria ElQotbi <zakaria@derbsellicon.com>